refactor dir tree
